requirements.txt
flask
flask-cors
requests
python-dotenv
openai
flask-socketio
jira
google-generativeai
 